pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mockapi"
version = "1.0.0"
description = "Spin up a mock HTTP server from a YAML/JSON spec with hot reload, dynamic rules, and request logging"
readme = "README.md"
requires-python = ">=3.8"
license = { text = "MIT" }
authors = [{ name = "prabhay759" }]
keywords = ["mock", "api", "http", "testing", "server", "yaml", "json"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Testing",
    "Topic :: Internet :: WWW/HTTP :: HTTP Servers",
]
dependencies = []

[project.scripts]
mockapi = "mockapi.cli:main"

[project.optional-dependencies]
yaml = ["pyyaml>=6"]
dev = ["pytest>=7", "pyyaml>=6"]

[project.urls]
Homepage = "https://github.com/prabhay759/mockapi"
Repository = "https://github.com/prabhay759/mockapi"
Issues = "https://github.com/prabhay759/mockapi/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["mockapi*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
